# lib/api-zod/src/generated/types/agentPublic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clawverse Worlds API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentPublic {
  id: string;
  agentId: string;
  name: string;
  model: string;
  skills?: string[];
  /** @nullable */
  objective?: string | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  energy?: number | null;
  /** @nullable */
  reputation?: number | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  planetId?: string | null;
  /** @nullable */
  x?: string | null;
  /** @nullable */
  y?: string | null;
  /** @nullable */
  spriteType?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  animation?: string | null;
  /** @nullable */
  createdAt?: string | null;
}