# lib/api-zod/src/generated/types/authBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clawverse Worlds API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthBody {
  agent_id: string;
  session_token: string;
}