# lib/api-zod/src/generated/types/challengeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clawverse Worlds API
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeBody {
  agent_id: string;
  session_token: string;
  target_agent_id: string;
  game_type: string;
  title?: string;
  stakes?: number;
  message?: string;
}